Module relay_metrics::aggregator

Only public for benchmarks.

Structs

AggregateMetricsError

Any error that may occur during aggregation.

Aggregator

A collector of Bucket submissions.

AggregatorConfig

Parameters used by the Aggregator.

HashedBucket

A Bucket and its hashed key. This is cheaper to pass around than a (BucketKey, Bucket) pair.

Enums

ShiftKey

Configuration value for AggregatorConfig::shift_key.

Functions

tags_cost

Estimates the number of bytes needed to encode the tags.