Struct relay_metrics::aggregator::Aggregator

pub struct Aggregator { /* private fields */ }

A collector of Bucket submissions.

Aggregation

Each metric is dispatched into the a Bucket depending on its project key (DSN), name, type, unit, tags and timestamp. The bucket timestamp is rounded to the precision declared by the bucket_interval field on the AggregatorConfig configuration.

Each bucket stores the accumulated value of submitted metrics:

Counter: Sum of values.

Distribution: A list of values.

Set: A unique set of hashed values.

Gauge: A summary of the reported values, see GaugeValue.

Conflicts

Metrics are uniquely identified by the combination of their name, type and unit. It is allowed to send metrics of different types and units under the same name. For example, sending a metric once as set and once as distribution will result in two actual metrics being recorded.

Implementations

impl Aggregator

pub fn name(&self) -> &str

Returns the name of the aggregator.

pub fn bucket_count(&self) -> usize

Returns the number of buckets in the aggregator.

pub fn totals_cost_exceeded(&self, max_total_cost: Option<usize>) -> bool

Returns true if the cost trackers value is larger than the given max cost.

pub fn into_buckets(self) -> Vec<Bucket>

Converts this aggregator into a vector of Bucket.

pub fn pop_flush_buckets( &mut self, force: bool ) -> HashMap<ProjectKey, Vec<HashedBucket>>

Pop and return the buckets that are eligible for flushing out according to bucket interval.

Note that this function is primarily intended for tests.

pub fn merge( &mut self, project_key: ProjectKey, bucket: Bucket, max_total_bucket_bytes: Option<usize> ) -> Result<(), AggregateMetricsError>

Merge a preaggregated bucket into this aggregator.

If no bucket exists for the given bucket key, a new bucket will be created.

pub fn merge_all<I>( &mut self, project_key: ProjectKey, buckets: I, max_total_bucket_bytes: Option<usize> )where I: IntoIterator<Item = Bucket>,

Merges all given buckets into this aggregator.

Buckets that do not exist yet will be created.

pub fn partition_buckets( &self, buckets: Vec<HashedBucket>, flush_partitions: Option<u64> ) -> BTreeMap<Option<u64>, Vec<Bucket>>

Split buckets into N logical partitions, determined by the bucket key.

pub fn new(config: AggregatorConfig) -> Self

Create a new aggregator.

pub fn named(name: String, config: AggregatorConfig) -> Self

Like Self::new, but with a provided name.