relay_metrics

Enum MetricUnit

pub enum MetricUnit {
    Duration(DurationUnit),
    Information(InformationUnit),
    Fraction(FractionUnit),
    Custom(CustomUnit),
    None,
}

The unit of measurement of a metric value.

Units augment metric values by giving them a magnitude and semantics. There are certain types of units that are subdivided in their precision, such as the DurationUnit for time measurements.

Units and their precisions are uniquely represented by a string identifier.

Variants

Duration(DurationUnit)

A time duration, defaulting to "millisecond".

Information(InformationUnit)

Size of information derived from bytes, defaulting to "byte".

Fraction(FractionUnit)

Fractions such as percentages, defaulting to "ratio".

Custom(CustomUnit)

User-defined units without built-in conversion or default.

None

Untyped value without a unit ("").

Implementations

impl MetricUnit

pub fn is_none(&self) -> bool

Returns true if the metric_unit is None.

pub fn as_str(&self) -> &str

Returns the string representation for this metric unit.