relay_metrics

Function normalize_bucket

pub fn normalize_bucket(bucket: &mut Bucket) -> Result<(), NormalizationError>

Normalizes a bucket.

The passed metric will have its name and tags normalized and tested for validity. Invalid characters in the metric name may be replaced, see relay_base_schema::metrics::try_normalize_metric_name.

Invalid tags are removed and tag keys are normalized, for example control characters are removed from tag keys.