Struct relay_metrics::Aggregator

pub struct Aggregator { /* fields omitted */ }

A collector of Metric submissions.

Aggregation

Each metric is dispatched into the a Bucket depending on its project key (DSN), name, type, unit, tags and timestamp. The bucket timestamp is rounded to the precision declared by the bucket_interval configuration.

Each bucket stores the accumulated value of submitted metrics:

Counter: Sum of values.

Distribution: A list of values.

Set: A unique set of hashed values.

Gauge: A summary of the reported values, see GaugeValue.

Conflicts

Metrics are uniquely identified by the combination of their name, type and unit. It is allowed to send metrics of different types and units under the same name. For example, sending a metric once as set and once as distribution will result in two actual metrics being recorded.

Flushing

Buckets are flushed to a receiver after their time window and a grace period have passed. Metrics with a recent timestamp are given a longer grace period than backdated metrics, which are flushed after a shorter debounce delay. See AggregatorConfig for configuration options.

Receivers must implement a handler for the FlushBuckets message:

use actix::prelude::*;
use relay_metrics::{Bucket, FlushBuckets};

struct BucketReceiver;

impl Actor for BucketReceiver {
    type Context = Context<Self>;
}

impl Handler<FlushBuckets> for BucketReceiver {
    type Result = Result<(), Vec<Bucket>>;

    fn handle(&mut self, msg: FlushBuckets, _ctx: &mut Self::Context) -> Self::Result {
        // Return `Ok` to consume the buckets or `Err` to send them back
        Err(msg.into_buckets())
    }
}

Implementations

impl Aggregator

pub fn new(config: AggregatorConfig, receiver: Recipient<FlushBuckets>) -> Self

Create a new aggregator and connect it to receiver.

The aggregator will flush a list of buckets to the receiver in regular intervals based on the given config.

pub fn insert(
&mut self,
project_key: ProjectKey,
metric: Metric
) -> Result<(), AggregateMetricsError>

Inserts a metric into the corresponding bucket in this aggregator.

If no bucket exists for the given bucket key, a new bucket will be created.

pub fn merge(
&mut self,
project_key: ProjectKey,
bucket: Bucket
) -> Result<(), AggregateMetricsError>

Merge a preaggregated bucket into this aggregator.

If no bucket exists for the given bucket key, a new bucket will be created.

pub fn merge_all<I>(
&mut self,
project_key: ProjectKey,
buckets: I
) -> Result<(), AggregateMetricsError> where
I: IntoIterator<Item = Bucket>,

Merges all given buckets into this aggregator.

Buckets that do not exist yet will be created.