Struct relay_metrics::DistributionValue

pub struct DistributionValue { /* private fields */ }

A distribution of values within a Bucket.

Distributions store a histogram of values. It allows to iterate both the distribution with iter and individual values with iter_values.

Based on individual reported values, distributions allow to query the maximum, minimum, or average of the reported values, as well as statistical quantiles.

Example

use relay_metrics::dist;

let mut dist = dist![1.0, 1.0, 1.0, 2.0];
dist.insert(5.0);
dist.insert_multi(3.0, 7);

Logically, this distribution is equivalent to this visualization:

value | count
1.0   | ***
2.0   | *
3.0   | *******
4.0   |
5.0   | *

Serialization

Distributions serialize as sorted lists of floating point values. The list contains one entry for each value in the distribution, including duplicates.

Implementations

impl DistributionValue

pub fn new() -> Self

Makes a new, empty DistributionValue.

Does not allocate anything on its own.

pub fn len(&self) -> u32

Returns the number of values in the map.

Examples

use relay_metrics::DistributionValue;

let mut dist = DistributionValue::new();
assert_eq!(dist.len(), 0);
dist.insert(1.0);
dist.insert(1.0);
assert_eq!(dist.len(), 2);

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

pub fn insert(&mut self, value: DistributionType) -> u32

Adds a value to the distribution.

Returns the number this value occurs in the distribution after inserting.

Examples

use relay_metrics::DistributionValue;

let mut dist = DistributionValue::new();
assert_eq!(dist.insert(1.0), 1);
assert_eq!(dist.insert(1.0), 2);
assert_eq!(dist.insert(2.0), 1);

pub fn insert_multi(&mut self, value: DistributionType, count: u32) -> u32

Adds a value multiple times to the distribution.

Returns the number this value occurs in the distribution after inserting.

Examples

use relay_metrics::DistributionValue;

let mut dist = DistributionValue::new();
assert_eq!(dist.insert_multi(1.0, 2), 2);
assert_eq!(dist.insert_multi(1.0, 3), 5);

pub fn contains(&self, value: impl Borrow<DistributionType>) -> bool

Returns true if the set contains a value.

Examples

use relay_metrics::dist;

let dist = dist![1.0];

assert_eq!(dist.contains(1.0), true);
assert_eq!(dist.contains(2.0), false);

pub fn get(&self, value: impl Borrow<DistributionType>) -> u32

Returns how often the given value occurs in the distribution.

Examples

use relay_metrics::dist;

let dist = dist![1.0, 1.0];

assert_eq!(dist.get(1.0), 2);
assert_eq!(dist.get(2.0), 0);

pub fn iter(&self) -> DistributionIter<'_>

Gets an iterator that visits unique values in the DistributionValue in ascending order.

The iterator yields pairs of values and their count in the distribution.

Examples

use relay_metrics::dist;

let dist = dist![2.0, 1.0, 3.0, 2.0];

let mut iter = dist.iter();
assert_eq!(iter.next(), Some((1.0, 1)));
assert_eq!(iter.next(), Some((2.0, 2)));
assert_eq!(iter.next(), Some((3.0, 1)));
assert_eq!(iter.next(), None);

pub fn iter_values(&self) -> DistributionValuesIter<'_>

Gets an iterator that visits the values in the DistributionValue in ascending order.

Examples

use relay_metrics::dist;

let dist = dist![2.0, 1.0, 3.0, 2.0];

let mut iter = dist.iter_values();
assert_eq!(iter.next(), Some(1.0));
assert_eq!(iter.next(), Some(2.0));
assert_eq!(iter.next(), Some(2.0));
assert_eq!(iter.next(), Some(3.0));
assert_eq!(iter.next(), None);