relay_metrics

Type Alias SetValue

pub type SetValue = BTreeSet<SetType>;

A set of unique values.

Set values can be specified as strings in the submission protocol. They are always hashed into a 32-bit value and the original value is dropped. If the submission protocol contains a 32-bit integer, it will be used directly, instead.

See the bucket docs for more information on set hashing.

Aliased Type

pub struct SetValue { /* private fields */ }