Crate relay_metrics

Metric protocol, aggregation and processing for Sentry.

Metrics are high-volume values sent from Sentry clients, integrations, or extracted from errors and transactions, that can be aggregated and queried over large time windows. As opposed to rich errors and transactions, metrics carry relatively little context information in tags with low cardinality.

Protocol

Clients submit metrics in a text-based protocol based on StatsD. A sample submission looks like this:

endpoint.response_time@ms:57|d|'1615889449|#route:user_index
endpoint.hits:1|c|'1615889449|#route:user_index

Structs

Metric
A single metric value representing the payload sent from clients.

ParseMetricError
An error returned by Metric::parse and Metric::parse_all.

ParseMetrics
Iterator over parsed metrics returned from Metric::parse_all.

UnixTimestamp
A unix timestap (full seconds elapsed since 1970).

Enums

DurationPrecision
Time duration units used in MetricUnit::Duration.

MetricType
The type of a metric, determining its aggregation and evaluation.

MetricUnit
The unit of measurement of a metric value.

MetricValue
The value of a metric.