Crate relay_metrics

Metric protocol, aggregation and processing for Sentry.

Metrics are high-volume values sent from Sentry clients, integrations, or extracted from errors and transactions, that can be aggregated and queried over large time windows. As opposed to rich errors and transactions, metrics carry relatively little context information in tags with low cardinality.

Protocol

Clients submit metrics in a text-based protocol based on StatsD. A sample submission looks like this:

endpoint.response_time@ms:57|d|#route:user_index
endpoint.hits:1|c|#route:user_index

The metric type is part of its signature just like the unit. Therefore, it is allowed to reuse a metric name for multiple metric types, which will result in multiple metrics being recorded.

Metric Envelopes

To send one or more metrics to Relay, the raw protocol is enclosed in an envelope item of type metrics:

{}
{"type": "metrics", "timestamp": 1615889440, ...}
endpoint.response_time@ms:57|d|#route:user_index
...

The timestamp in the item header is used to send backdated metrics. If it is omitted, the received time of the envelope is assumed.

Aggregation

Relay accumulates all metrics in time buckets before sending them onwards. Aggregation is handled by the Aggregator, which should be created once per Project Key (DSN). It flushes aggregates in regular intervals, either shortly after their original time window has passed or with a debounce delay for backdated submissions.

Warning: With chained Relays submission delays accumulate.

Aggregate buckets are encoded in JSON with the following schema:

[
  {
    "name": "endpoint.response_time",
    "unit": "ms",
    "value": [36, 49, 57, 68],
    "type": "d",
    "timestamp": 1615889440,
    "tags": {
      "route": "user_index"
    }
  },
  {
    "name": "endpoint.hits",
    "value": 4,
    "type": "c",
    "timestamp": 1615889440,
    "tags": {
      "route": "user_index"
    }
  }
]

Ingestion

Processing Relays write aggregate buckets into the ingestion Kafka stream. The schema is similar to the aggregation payload, with the addition of scoping information:

[
  {
    "org_id": 1,
    "project_id": 42,
    "name": "endpoint.response_time",
    "unit": "ms",
    "value": [36, 49, 57, 68],
    "type": "d",
    "timestamp": 1615889440,
    "tags": {
      "route": "user_index"
    }
  },
  {
    "org_id": 1,
    "project_id": 42,
    "name": "endpoint.hits",
    "value": 4,
    "type": "c",
    "timestamp": 1615889440,
    "tags": {
      "route": "user_index"
    }
  }
]

Macros

dist
Creates a DistributionValue containing the given arguments.

Structs

AggregateMetricsError
Any error that may occur during aggregation.

Aggregator
A collector of Metric submissions. Every project key (DSN) has one Aggregator.

AggregatorConfig
Parameters used by the Aggregator.

Bucket
An aggregation of metric values by the Aggregator.

DistributionIter
An iterator over distribution entries in a DistributionValue.

DistributionValue
A distribution of values within a Bucket.

DistributionValuesIter
An iterator over all individual values in a DistributionValue.

FlushBuckets
A message containing a vector of buckets to be flushed.

GaugeValue
A snapshot of values within a Bucket.

InsertMetrics
A message containing a list of Metrics to be inserted into the aggregator.

MergeBuckets
A message containing a list of Buckets to be inserted into the aggregator.

Metric
A single metric value representing the payload sent from clients.

ParseBucketError
Error returned when parsing or serializing a Bucket.

ParseMetricError
An error returned by Metric::parse and Metric::parse_all.

ParseMetrics
Iterator over parsed metrics returned from Metric::parse_all.

UnixTimestamp
A unix timestamp (full seconds elapsed since 1970-01-01 00:00 UTC).

Enums

BucketValue
The aggregated value of a metric bucket.

DurationPrecision
Time duration units used in MetricUnit::Duration.

MetricType
The type of a MetricValue, determining its aggregation and evaluation.

MetricUnit
The unit of measurement of a metric value.

MetricValue
The typed value of a metric.