Struct relay_metrics::Metric

pub struct Metric {
    pub name: String,
    pub value: MetricValue,
    pub timestamp: UnixTimestamp,
    pub tags: BTreeMap<String, String>,
}

A single metric value representing the payload sent from clients.

As opposed to bucketed metric aggregations, this single metrics always represent a single submission and cannot store multiple values.

See the crate documentation for general information on Metrics.

Fields
name: String

The metric resource identifier, in short MRI.

MRIs follow three core principles:

Robustness: Metrics must be addressed via a stable identifier. During ingestion in Relay and Snuba, metrics are preaggregated and bucketed based on this identifier, so it cannot change over time without breaking bucketing.

Uniqueness: The identifier for metrics must be unique across variations of units and metric types, within and across use cases, as well as between projects and organizations.

Abstraction: The user-facing product changes its terminology over time, and splits concepts into smaller parts. The internal metric identifiers must abstract from that, and offer sufficient granularity to allow for such changes.

MRIs have the format <type>:<ns>/<name>@<unit>, comprising the following components:

Type: counter (c), set (s), distribution (d), gauge (g), and evaluated (e) for derived numeric metrics (the latter is a pure query-time construct and is not relevant to Relay or ingestion). See MetricType.

Namespace: Identifying the product entity and use case affiliation of the metric. See MetricNamespace.

Name: The display name of the metric in the allowed character set.

Unit: The verbatim unit name. See MetricUnit.

Parsing a metric (or set of metrics) should not fail hard if the MRI is invalid, so this is typed as string. Later in the metrics aggregator, the MRI is parsed using MetricResourceIdentifier and validated for invalid characters as well. MetricResourceIdentifier is also used in the kafka producer to route certain namespaces to certain topics.

value: MetricValue

The value of the metric.

Distributions and counters require numeric values which can either be integral or floating point. In contrast, sets and gauges can store any unique value including custom strings.

timestamp: UnixTimestamp

The timestamp for this metric value.

tags: BTreeMap<String, String>

A list of tags adding dimensions to the metric for filtering and aggregation.

Tags are preceded with a hash # and specified in a comma (,) separated list. Each tag can either be a tag name, or a name:value combination. For tags with missing values, an empty "" value is assumed.

Tags are optional and can be omitted.

Implementations

impl Metric

pub fn new_mri( namespace: MetricNamespace, name: impl AsRef<str>, unit: MetricUnit, value: MetricValue, timestamp: UnixTimestamp, tags: BTreeMap<String, String> ) -> Self

Creates a new metric using the MRI naming format.

See Metric::name.

MRI is the metric resource identifier in the format <type>:<ns>/<name>@<unit>. This name ensures that just the name determines correct bucketing of metrics with name collisions.