Struct relay_metrics::Metric

pub struct Metric {
    pub name: String,
    pub unit: MetricUnit,
    pub value: MetricValue,
    pub ty: MetricType,
    pub timestamp: UnixTimestamp,
    pub tags: BTreeMap<String, String>,
}

A single metric value representing the payload sent from clients.

As opposed to bucketed metric aggregations, this single metrics always represent a single submission and cannot store multiple values.

See the crate documentation for general information on Metrics.

Submission Protocol

<name>[@unit]:<value>|<type>|'<timestamp>|#<tag_key>:<tag_value>,<tag>

See the field documentation on this struct for more information on the components. An example submission looks like this:

endpoint.response_time@ms:57|d|'1615889449|#route:user_index
endpoint.hits:1|c|'1615889449|#route:user_index

To parse a submission payload, use Metric::parse_all.

JSON representation

In addition to the submission protocol, metrics can be represented as structured data in JSON. Field values are the same with a single exception: The timestamp is required in JSON notation.

{
  "name": "endpoint.response_time",
  "unit": "ms",
  "value": 57,
  "type": "d",
  "timestamp": 1615889449,
  "tags": {
    "route": "user_index"
  }
}

Fields
name: String

The name of the metric without its unit.

Metric names can consist of ASCII alphanumerics, underscores and periods.

unit: MetricUnit

The unit of the metric value.

Units augment metric values by giving them a magnitude and semantics. There are certain types of units that are subdivided in their precision, such as the DurationPrecision for time measurements.

The unit can be omitted and defaults to MetricUnit::None.

value: MetricValue

The value of the metric.

Distributions and counters require numeric values which can either be integral or floating point. In contrast, sets and gauges can store any unique value including custom strings.

ty: MetricType

The type of the metric, determining its aggregation and evaluation.

timestamp: UnixTimestamp

The timestamp for this metric value.

In the SDK protocol, timestamps are optional and preceded with a single quote '. Supply a default timestamp to Metric::parse or Metric::parse_all to insert a default timestamp.

tags: BTreeMap<String, String>

A list of tags adding dimensions to the metric for filtering and aggregation.

Tags are preceded with a hash # and specified in a comma (,) separated list. Each tag can either be a tag name, or a name:value combination. For tags with missing values, an empty "" value is assumed.

Tags are optional and can be omitted.

Implementations

impl Metric

pub fn parse(
slice: &[u8],
timestamp: UnixTimestamp
) -> Result<Self, ParseMetricError>

Parses a single metric value from the raw protocol.

See the Metric for more information on the protocol.

Example

use relay_metrics::{Metric, UnixTimestamp};

let metric = Metric::parse(b"response_time@ms:57|d", UnixTimestamp::now())
    .expect("metric should parse");

pub fn parse_all(slice: &[u8], timestamp: UnixTimestamp) -> ParseMetrics<'_>

Parses a set of metric values from the raw protocol.

Returns a metric result for each line in slice, ignoring empty lines. Both UNIX newlines (\n) and Windows newlines (\r\n) are supported.

It is possible to continue consuming the iterator after Err is yielded.

See the Metric for more information on the protocol.

Example

use relay_metrics::{Metric, UnixTimestamp};

let data = br#"
endpoint.response_time@ms:57|d
endpoint.hits:1|c
"#;

for metric_result in Metric::parse_all(data, UnixTimestamp::now()) {
    let metric = metric_result.expect("metric should parse");
    println!("Metric {}: {}", metric.name, metric.value);
}

pub fn serialize(&self) -> String

Serializes the metric to the raw protocol.

See the Metric for more information on the protocol.

Example

use std::collections::BTreeMap;
use relay_metrics::{Metric, MetricUnit, MetricValue, MetricType, UnixTimestamp};

let metric = Metric {
    name: "hits".to_owned(),
    unit: MetricUnit::None,
    value: MetricValue::Integer(1),
    ty: MetricType::Counter,
    timestamp: UnixTimestamp::from_secs(1615889449),
    tags: BTreeMap::new(),
};

assert_eq!(metric.serialize(), "hits:1|c|'1615889449");